add a few more tests
